requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django
django-bootstrap4
idna==2.8
pytz==2019.1
requests==2.21.0
urllib3==1.24.3
django-ckeditor